Fix: Applying multiple discounts results in incorrect discountTotal (#2107)
Fix: When multiple Discounts apply to the same CartLine, the
subTotalDiscount and discountBreakdown contain both discounts, but the
discountTotal of the Cart and CartLine only contained the value of the
last Discount.

This inconsistency resulted in values not correctly adding up in the
Admin Panel Order Detail and other places.

This commit fixed both the discountTotal of the Cart and the CartLine to
match the Cart's discountBreakdown, subTotalDiscounted & total.

---------

Co-authored-by: tim <tim@direct-effect.nl>
Co-authored-by: Author <actions@github.com>
Co-authored-by: Alec Ritson <hello@itsalec.co.uk>