Allow using locales when creating or editing languages in the admin panel (#2026)
Allow entering a maximum of 5 characters for a language code instead of
2, so that a locale can be used. (eg en_GB or nl-BE).

There are multiple use cases for this:
- Words and spelling can differ between different versions of the same
language.
- Multiple domain names may be used, and different descriptions can be
useful to avoid a duplicate content SEO penalty.

I believe this is the only change necessary to support using locales.
The database column already allows long language codes, and I don't see
other places where this could create an issue.

Co-authored-by: tim <tim@direct-effect.nl>