Add product link to order lines (#2178)
It might be useful for store owners to be able to click through to a
product when viewing order lines. There is a check to ensure we only try
to link if the purchasable item is a variant, we may want to look to
utlisie extensions in the future to accomodate different purchasables
stores might have.

---------

Co-authored-by: Author <actions@github.com>