Refactor order reference generator (#2175)
This PR changes the way order references are generated. Currently the
format YYYY-MM-XXXXXX creates difficulty when attemping to determine the
next reference in the sequence due to being a string.

The decision was to move away from any sort of queries when determining
the next reference and instead just use the ID of the order and allow
developers to customise certain formatting options within config, so in
most cases there should be no need to use a custom generator for small
changes.

---------

Co-authored-by: Glenn Jacobs <glenn@neondigital.co.uk>
Co-authored-by: Author <actions@github.com>