Fix: The "total" column of the shipping order-lines in the database doesn't include taxes (#2034)
When prices are stored exclusive of tax, the subtotal should exclude
taxes, and the total should include taxes.

The issue occurs because the CalculateTax pipeline adds the tax to the
cart's shippingTotal, but not to the shippingAddress' shippingTotal. The
address' shippingTotal is then used by the CreateShippingLine pipeline
to create the record in the lunar_order_lines table.

Screenshot of the lunar_order_lines table, filtered to only show the
shipping lines:


![image](https://github.com/user-attachments/assets/7a1c85ac-8848-438d-8991-e63952277a7f)

Also: The TestTaxDriver doesn't take into account whether prices are
stored inclusive of tax or not, which results in it calculating the
wrong taxBreakdown if they are.

Co-authored-by: tim <tim@direct-effect.nl>
Co-authored-by: Glenn Jacobs <glenn@neondigital.co.uk>