Fix line fetching in order creation (#1956)
I'm updating the line fetching in
`Lunar\Pipelines\Order\Creation\CreateOrderLines` and
`Lunar\Pipelines\Order\Creation\MapDiscountBreakdown` to match the one
used in `Lunar\Actions\Carts\GetExistingCartLine` to see if the line
meta matches as well.

This PR will help be able to handle products when similar products are
added instead of merged and quantity updated.

---------

Co-authored-by: Author <actions@github.com>
Co-authored-by: Glenn Jacobs <glenn@neondigital.co.uk>