Fix validation on prices in Postgres (#2234)
Closes #1974 

The pricing unique validation wasn't working in any databases, and made
Postgres return an error.

The unique validation isn't required for customer group pricing, and
it's only really a helper for price breaks, nothing will break if
someone decides to use the same price for two tiers and it might
actually be useful in some scenarios.

---------

Co-authored-by: Author <actions@github.com>