Fix error in collection deletion (#2110)
When deleting a collection that has child collections in the admin
panel, it will error. I believe this is because the child-collections
attempt to be removed by nestedset, but their associated customer groups
etc are not being removed.

The simplest and safest approach might be to simply not allow the
deletion of collections if they have child collections.

This PR introduces a quick solution for that, but there may be a nicer
way.

---------

Co-authored-by: Author <actions@github.com>