Update CollectionResource.php (#2249)
We have over 15,000 collections in the main catalog. The problem with
**ancestors** is that it shows the linear collections that are on the
same level. This makes the breadcrumb 15,000+ collection names. Using
**CHILDREN** only shows the children of the collection that you clicked
into. The page loads faster and the title bar with the breadcrumb is
more readable.

Please describe your Pull Request as best as possible, explaining what
it provides and why it is of value, referencing any related Issues.

Try to include the following in your Pull Request, where applicable...

- Documentation updates
- Automated tests


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

* **Bug Fixes**
* Updated breadcrumb trail logic to display direct children of a
collection instead of its ancestors.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

Co-authored-by: Alec Ritson <hello@itsalec.co.uk>