Fix admin activity log feed rendering (#2314)
This PR fixes 2 issues when rendering the 'Activity Log' component on
the order screen in the admin.

The first issue is visible when using Filament Notifications. The
z-index of the comment form in the activity log is set to 20 which
brings it in front of the Filament notifications sidebar when it is
opened.

The second issue is when using dark mode. The date heading in the
activity log is using the 'text-gray-300' Tailwind class, however this
class isn't generated so the 'date' heading in the activity log is black
and cannot be read.

---------

Co-authored-by: Glenn Jacobs <glenn@neondigital.co.uk>