Fix: Allow the discount to be applied if no condition coupon exists regardless of the coupon code on the cart (#2300)
Allow the discounts to be applied if no condition coupon exists
regardless of cart coupon. This way discounts that doesn't have a coupon
requirement can be added and not be removed because of another coupon
discount that is applied.

Current behaviour:
if a discount that has a coupon requirement is applied. and other
discounts that doesn't have a coupon requirements are applied as well.
the coupon code discount causes the other discounts not to apply to the
cart at all because of a wrong $validCoupon comparison logic.

Expected behaviour:
discounts should be applied if there is no coupon code requirement
regardless of the coupon code set on the cart. this way coupon discounts
doesn't simply override all other discounts.

---------

Co-authored-by: Glenn Jacobs <glenn@neondigital.co.uk>