Fix discount conditions on collections (#2345)
This PR builds on the work done in
https://github.com/lunarphp/lunar/pull/2250

It:
1. Adds a brand check 
2. Seperates collections into its own relation manager to allow it to
continue to use collection_discount table

It's breaking, but for v2 we should consolidate everything into the
discountable table - I missed these when I did this initial piece of
work. It would make the UI much more streamlined and from a data point
of view reduce the redundancy of the tables.

---------

Co-authored-by: fredmarmillod <fmarmillod@gmail.com>
Co-authored-by: Author <actions@github.com>
Co-authored-by: Frédéric Marmillod <9134609+fredmarmillod@users.noreply.github.com>
Co-authored-by: Glenn Jacobs <glenn@neondigital.co.uk>